+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.7.4 (2018-08-16)
+
+### Added 
+
+- `VideoConvertParams::withPass(int $passNumber)` to indicate ffmpeg pass number
+
+### Documentation
+
+- Multipass example fixed using `VideoConvertParams::withPass()`.
+
+
 ## 0.7.3 (2018-08-15)
 
 ### Added